Bug#35049440: Mysqld assert in Item_typecast_signed::val_int
Problem is in Item_field::replace_equal_field where we try to
replace fields with equivalent ones in order to push some of
these predicates down, we do not take into consideration
nullability, so some non-nullable fields were replaced with nullable
ones.
The fix was to add a function called allow_replacement which determines
if a non-nullable field can be replaced with a nullable one
(UNKNOWN results can be treated as false), otherwise skip such changes.

Change-Id: Id4370babd6d6e29e36a05ed46e4ca951507670d6