Bug#34199339 Replica fail for keyless table using BIT and HASH_SCAN
Problem:

When replicating a NDB table with BIT columns and with no explicit primary key (NDB
hidden PK), and the search algorithm of replica rows is HASH_SCAN,
replication can stop with a failure on the SQL applier thread.

Analysis:

The HASH_SCAN algorithm will compute a hash value using the
`null_flags` of a table (info of nullable columns and the first
high-bit of BIT columns) and the BI field data. This hash value
is then used populate a hash table, after which the same computation
will take place using the data retrieved from NDB. If the hash
matches, a row is found and the event is applied.

It was observed that the first hash never matched the hash of data
from NDB. The value that was making the hash have wrong value was the
`null_flags` of the table. The intricacies of this value is storing
the null bits of the columns as well as the first 0-6 uneven bits of
BIT columns (the extra unaligned bits).

Solution:

It was thus found that the unpack of the record, at the ndbcluster
plugin level, trusted that the null bits were already set so it did
nothing, i.e., did not reset the high bits on the destination buffer.
Now, on the unpack of the NDB record, the ndbcluster plugin writes the
value of 0 if a field is NULL, forcing the uneven high bits to be 0.

Change-Id: I25e7ba2df8420a1495f317eb2e2c472ad98dd91e