Bug#35943415 mysql_native_password account reused on bootstrap if exists.
When the Router is bootstrapped over an existing configuration it
will fetch the existing user's credentialis from the keyring and reuse
this user account. This happens even if the existing user is using
the deprecated mysql_native_password authentication plugin.

This patch changes the behavior so that the Router is checking what
is the authentication plugin for the Router user. If it is
mysql_native_password it will try to change it to the Cluster's
default authentication plugin. It will only do that if there is
a single user/host entry for the Router's user in the mysql.user
table. If there is more than one host patter it will only give
a warning, advising user to change the authentication plugin manually.

Change-Id: I9477cca244076488cb4da78307c6a10db646642b