Bug#35925503 Message "Metadata: Failed to submit table 'mysql.ndb_apply_status' for synchronization" is submitted every minute
Problem:
On MySQL Cluster 8.0.34, messages like below are submitted to the error
log every minute:

  > Metadata: Failed to submit table 'mysql.ndb_apply_status' for
synchronization

This is essentially useless, and fills the error log unnecessarily.

Analysis:
The mysql.ndb_apply_status is a util table managed by the binlog thread,
as such there is no need to check if it has changed.

Solution:
Remove table from list of tables in DD, this is done in a similar fashion
to how it's removed from the list of tables in NDB. This  means that the
table will not be submitted for change.

Change-Id: I7af9579bcd98b0e7db5b2acccf0b0a7df89a7265