Bug#32764586 : MEMORY LEAK WHEN ACCESSING VIEW
Description:
------------
Each SELECT on a view causes a small memory leak on the
Windows platform.

Analysis:
---------
In case of select view Security_context(sctx) is allocated
in the MEM_ROOT. Memory allocated by the members of sctx is
not freed until the MEM_ROOT goes out of scope as a result
memory keeps on accumulating. Security_context employed a
mechanism logout() to avoid such memory accumulation but for
it to work correctly, we need to own the life cycle of the
sctx members keeping in mind that it is a frequently
accessed code path so that we avoid frequent memory
allocations/deallocations.

DB_restrictions is owned by the Security_context(sctx). The
former has an associative container std::unordered_map
(m_restrictions) member that is allocated on the stack. We
need to clear the memory allocated to this member at the
time of logout(). We had a similar memory growth reported
in Bug#31919448. At that time we used the temporary object
and swap() trick to free the content of m_restrictions.
Unfortunately, that trick worked on Linux but not on
Windows. This is because constructors of associative
containers can throw which means compilers are free to
allocate memory on the heap in the constructor of
associative containers. As it turned out to be the case on
Windows, std::unordered_map allocates the memory in the
heap. It is not cleared unless the object is destroyed. That
means the swap() trick would not work in such cases because
memory allocated by the temporary object is transferred to
the m_restrictions hence memory growth is still seen.

We have the following options to fix this situation :

- Refactor the code in the select view area that may have
  a wider repercussions, hence it is not the preferred
  solution.

- Create the m_restrictions map on the heap but not on the
  stack. This is the preferred option.

Fix:
----
- Change m_restrictions object creation from stack to heap.
- Allocate the memory to m_restriction only if that is
  required.
- Handle the unsafe usage of the partial_revokes APIs.
- Got rid of the swap() trick in DB_restrictions::clear().
- Removed is_not_empty() as we already have is_empty().

Change-Id: Iee828909d23991db86227f9664eb7f130a3d6d56