Bug#33800633 startChangeNeighbour problem
When setting up a new set of neighbour transporters, the old set
of neighbours may still have pending data awaiting to be sent.

In order to ensure that available data on these transporters
will still be sent, we need to ensure that the becoming non-neighbour
transporters are inserted into the list of transporter we need
to send on.

Patch enhance startChangeNeighbour() such that it will check
the old set of neighbour transporters if any has 'm_data_available'
and use insert_trp() to insert them into the list of non-neighbour
transporters.

Note that we now also need to clear the m_neighbour_trp flag before
doing such inserts, else insert_trp() would not have inserted the
TrpId into the non-neighbour list.

Patch also redeclares some local variables refering a 'transporter id'
from an Uint32 to a TrpId.

A few asserts are also added to ensure the consistency of the
Transporter list structures.

Change-Id: I87bb539868ef33eb19e37ed90b30f1a3e1e07680