BUG - Crash when while executing query MySQL Server disconnects MRS plugin
Description
===========
After server disconnect, the mysqlsession class throws an exception which
leads to return of MySQLSession to the connection cache.
The State of the MySQLSession is invalid, and after trying to reuse the
connection the state is not checked.

Fix
===
Before caching the connection, check last error and if its
a client error (CR_*) then drop such connection.