BUG#35949017 Schema dist setup lockup
Problem:
Schema event handler hangs waiting for GSL to become available when
trying to setup util tables.

Analysis:
The schema event handler will try to setup the util tables when the
physical tables have been dropped from NDB or connection lost otherwise.
This setup check is intended to be done periodically every second but
since the first step is to acquire the GSL the check will lockup waiting
for the GSL. During this time the schema event handler and thus the
binlog thread will not be actively handling other events.

Solution:
Use try lock when attempting to acquire the GSL, thus causing another
setup check attempt to be made at a later time if GSl is not available.

This change has the effect that check setup is active during a slightly
longer then before, this means the util_tables_drop test has to be
modified to wait for all servers completing the setup before continuing
with next step in test. Previoulsy the test only waited for servers that
would be used for subsequent SQL queries.

Change-Id: Ie8e90b65781d94737eafc0e2ae0b6f3c18599f1d