WL#15822 Log progress information while rolling back large transactions
When rolling back long-running transactions, information-level messages
are periodically emitted to error log while the rollback is in progress.
The first message, which also contains additional diagnostic
information, is logged after 30s and subsequent messages are logged
every 30s. Progress information is provided based on number of rows to
undo.

Change-Id: Ie4cabd9c73ff399199d10894a16eac2d675e02df