Bug#35932118 	innodb.redo_log_archive_01 fails with log0write.cc:2115:ib::fatal
The `os_innodb_umask` is global variable, it can't be modified by a thread that wants to create a file with different UNIX access mode, as it will modify it for all threads (not mentioning UndefinedBehavior).
The `os_file_set_umask` is supposed to be called only once, at InnoDB initialization.
However, after `Bug #29472125 NEED OS_FILE_GET_UMASK()` it was made possible to modify it and `WL#12009 - Redo-log Archiving` used it to modify it.
Fix:
- `os_file_get_umask()` is removed
- `os_file_set_umask()` is modified so it can be called only once
- Unix-only `os_file_create_simple_no_error_handling_with_umask()` is added to allow specifying umask param.

Change-Id: I78f169cfb99704e031ea1ff758b970cd8c73240d