Bug#35928350 Contribution: NDB STORED USER fails with replication filter
Problem:
The MySQL Server fails when replicating "GRANT NDB_STORED_USER ..." with
replication filter turned on. This occurs since the replication filter
causes all non-updating queries to return an error - due to the
assumption that only changes need to be replicated.

Analysis:
To handle the GRANT statement a `SELECT ... FROM
information_schema.user_privileges` is used, this is a non-updating
query and thus triggers an error in the MySQL Server when replication
filters are in use.

Solution:
Install an empty replication filter while running distributed privilege
queries, thus sucessfully getting a result returned.

Thanks to Mikael Ronström for providing the steps necessary how
to reproduce this problem.

Contributed by: Mikael Ronström

Change-Id: I2c171f64b7776ac2410d2d918f591b16bc8c1a04