BUG#27934653 60% OVERHEAD FROM PERFORMANCE SCHEMA INSTRUMENTATION OF STORED FUNCTIONS
Problem
=======

When executing a stored program,
the performance schema can create some overhead.

Root cause
==========

The overhead varies depending on the stored program code.

For instructions invoking SQL statements (statement/sp/stmt),
the overhead is similar to the overhead the same statement
has when executed from a top level query.
This part is working as expected.

For instructions invoking the stored procedure "micro instructions",
which are:
- statement/sp/set
- statement/sp/set_trigger_field
- statement/sp/jump
- statement/sp/jump_if_not
- statement/sp/freturn
- statement/sp/hpush_jump
- statement/sp/hpop
- statement/sp/hreturn
- statement/sp/cpush
- statement/sp/cpop
- statement/sp/copen
- statement/sp/cclose
- statement/sp/cfetch
- statement/sp/error
- statement/sp/set_case_expr

the overhead of:
- creating a sub statement instrumentation
- starting a wait timer
- starting a cpu timer
- stoping a cpu timer
- storing a wait timer
- measuring memory consumption
- aggregating all statements attributes

can be very expensive compared to the
actual micro statement execution.

For example, for simple:
  SET i = i + 1;
there is no point in collecting
the full statement instrumentation.

Fix
===

Introduced new flags, PSI_FLAG_DISABLED and PSI_FLAG_UNTIMED,
to allow a statement instrument to default to
ENABLED='NO' and TIMED='NO' respectively.

Modified all the stored procedure micro instructions,
'statement/sp/%' to be disabled by default.

The instrument 'statement/sp/stmt' is unchanged.

Also, provided documentation for stored procedures instruments.

Test
====

Tested manually that stored procedure instruments changed
are ENABLED='NO' and TIMED='NO' by default.

The test can not be automated, because the test suite is configured with:
  loose-performance-schema-instrument='%=ON'
which overrides compile time defaults.

Tested manually with the levenshtein.sql test case
provided in the bug report that the performance issue
is resolved.

Change-Id: Id604940c58dcb27cb5064d71471c3789a1b828fc