Bug #35917292 Add support for building Oracle Linux RPM with PGO
Follow-up correction: The variant of the fix that was pushed to trunk
was adopted for 8.2.0 which was about to be respun.

However, it is not correct for trunk since the cmake command line has
changed. This patch fixes that by making the PGO rebuild command
identical to the normal build.

Change-Id: Ib735bb248acc299fbd8eaa63c706bb8e9e5f0e20