Bug#35877452: mysqld crash at LogicalOrderings::AddHomogenizedOrderingIfPossible
A query with redundant elements in the ORDER BY clause could fail if
the query optimization took so long time that the secondary engine
requested a restart of the optimization with a different set of
parameters to restrict the search space.

BuildInterestingOrders() got confused because Query_block::order_list
was in an inconsistent state when the optimization was restarted. It
was inconsistent because the optimizer had modified the intrusive list
pointers in JOIN::order to remove redundant elements. JOIN::order
shares some data with Query_block::order_list, so some of the
underlying data of Query_block::order_list was modified as a result
without the parent object's knowledge, and the parent object became
inconsistent.

Query_block::order_list is restored before the next execution of a
prepared statement for this exact reason; see
Query_block::restore_cmd_properties() and the doxygen comment for
Query_block::order_list. But it is not restored when the optimization
is restarted. Since it's not safe to access Query_block::order_list
after such modifications, the optimizer should use JOIN::order
instead.

Fixed by making BuildInterestingOrders() inspect JOIN::order instead
of Query_block::order_list.

For completeness and consistency, the patch also replaces usages of
Query_block::group_list in the hypergraph optimizer with corresponding
data structures in JOIN, even though the hypergraph optimizer doesn't
currently mutate the group list in a similar way.
BuildInterestingOrders() now uses JOIN::group_list instead of
Query_block::group_list, and EstimateAggregateRows() uses
JOIN::group_fields. (EstimateAggregateRows() already used
JOIN::group_fields for some cases, but only when called by the old
optimizer. It could not use it for the hypergraph optimizer, because
the hypergraph optimizer didn't populate group_fields until after
EstimateAggregateRows(). The patch moved the hypergraph optimizer's
call to make_group_fields() a little earlier so that
EstimateAggregateRows() could use the same code for both optimizers.)

Change-Id: I4790482416e8935b7918cccd36f316d88b1d5700