Bug#35647759: crash with 'install component' and set global  with subquery
When a function is itemized the flags for the arguments are collected
first and *THEN* the arguments are itemized.
This doesn't account for certain Item itemizations that would replace
the item.
For these the flags aren't aggregated and INSTALL COMPONENT fails to
correctly detect and report the unsupported SET value expression types.

Fixed by itemizing the argument first and then collecting the flags.

Change-Id: I7a1ab3e6283b0af23d015ec8e54d7bb71e28181e