Bug#27910375 JOIN::TABLE COUNT IS WRONG
As it turns out, the code is not wrong, but the comment explaining the
extra +1 is. The +1 is not contingent only on having windowing, there
are other reason we may want to materialize the result set to a tmp
table as well, among them ROLLUP and full text search, see logic of
setting JOIN::need_tmp_before_win.

So, the patch is only a comment change.

Change-Id: I7d80b2fa8baa927c3fcd9c0f255c7bed5c26e647