Bug#34724344: MySQLd crashing on START GROUP_REPLICATION
Issue:
======
When the asynchronous replication connection failover is enabled and below
happens:
i.  The receiver tries to reconnect to alternate source.
ii. The receiver IO thread is killed.

then,
the mysql_close() gets called twice for the same MYSQL object resulting in mysqld
crash.

The MYSQL object when released first in handle_slave_io() function in asynchronous
replication failover re-connection code part, it was not set to nullptr, due to
which it was called again when IO thread exits to release MYSQL object resulting
in crash.

Solution:
=========
The MYSQL object when released in handle_slave_io() function in asynchronous
replication failover re-connection code part, it should be also set to nullptr.

Change-Id: I50d00fa10894200516e26e4b97fbc772daba891f