Bug#35839977 testNodeRestart -n Bug34216 fails randomly
Test executes a number of updates against a table it created.
Some of the updates are executed as 'NoCommit', with a final 'Commit'
for the entire transaction.

ERROR_INSERT is used to crash the node on the commit (as apposed
to the intermediate execute-NoCommit's)

Intention of test case is to check for regression against Bug#34216,
where: 'During TC-take-over (during NF) commit messages can come
out of order to TUP' - Thus presumable hitting the ERROR_INSERT
when execute-commit. Such failures are now randomly detected
by the failing test case.

Root cause is that there is other background activity going on on
the data nodes as well. In particular the update operations may
trigger ndb_index_stat updates performing READ operations on
the table being updated, as well as accessing the system tables.

Patch enhance the ERROR_INSERT 5048/5049 code to also require that
it is a ZUPDATE being commited, as well as that the table is
an UserTable.

Change-Id: I4061fa4521b8f670b3783a3aa0b6256bca15a7d0