Bug#35945822: SIGABRT Mysqld crash from operator() at sql_executor.cc:4399
A query using GROUP BY WITH ROLLUP fails because it is not able to map
an expression in the HAVING clause to the corresponding grouping
column.

The reason why it doesn't find the grouping column, is that Item::eq()
gets confused by Item_cache items being added to one of the items
being compared for equivalence, but not to the other. Item::eq() does
try to see through Item_cache and other wrappers, but the current
logic works only if the wrappers are applied in a given order. For
example, if a cache is on top of a rollup group wrapper, it is able to
see through both the cache and the rollup group wrapper. However, if
the rollup group wrapper is on top of the cache, it sees through the
rollup group wrapper, but not through the cache.

Fixed by making the unwrapping logic keep going as long as there is
something more to unwrap.

Change-Id: I0359f8a56f23ea93c80d8256f02ddbdb582e7f69