Bug#34551954 Slave SQL Thread crashes the instance when create user
Problem:
Crash in replication applier when handling NDB synchronized privilieges.

Analysis:
Fatal failure occurs when query return code indicates sucess but no
result set is available. Such problem scenario can be seen in the
replication applier, for example in problem described in BUG#35928350.

Solution:
Handle the case when expected result set is not available by logging an
error message and returning failure. This should make for a more stable
behaviour when the queries does not return the expected result set.

Change-Id: I4623ff8be5e59cbc3edcbd1627cbedda6dbab5ce