Bug#35968017 Unknown character set '' with c/j and rwsplitting
When connector/j is used to connect to the rw-splitting port, the
connection fails and router logs:

  Unknown character set ''

The error is triggered by connector/j as it runs:

  SET character_set_results = NULL

to disable conversion of the character sets from server to client, but
the session-tracker reports the NULL as '' to the router. Replaying
the SET character_set_results = "" leads to above error as empty-string
isn't a known character-set name.

Change
======

- store NULL instead of "" when the session-tracker reports
  { character_set_results: "" }

Change-Id: Idcfc9b7668e034e0f0fd3efedac810ad87035ff3