Bug#35899768 Api failure handling during node restart leaves subscriptions behind
During a data node restart, the starting node asks nodegroup peers
to copy the details of the current event subscribers to the starting
node, so that it can take over responsibility for part of the event
forwarding and buffering work when it is started.

This implies that the starting node must also monitor for subscribing
API node failures from this point forward.

As the copy of subscribers happens before the API nodes are directly
connected to the starting node, it is necessary for the starting data
node to handle API failures even when those APIs are not directly
connected during restarts.

This was already handled to some extent in QMGR, but appears to have
regressed over time.

This could result in e.g. :
 - Leaked subscribers at SUMA
 - Wasted effort forwarding events to disconnected subscribers
 - Wasted effort forwarding events to reconnected subscribers
 - Crashes / undefined behaviour if reconnected subscribers reuse
   NdbApi object identitied.

test_event_mysqld is created to give some coverage of the MySQLD
ha_ndbcluster plugin stack event consumption behaviour during data
node restarts and MySQLD asynchronous disconnects.

  test_event_mysqld -n MySQLDEventsRestarts
  test_event_mysqld -n MySQLDEventsDisconnects
  test_event_mysqld -n MySQLDEventsRestartsDisconnects

The third testase is invoked from a new MTR testcase to give simple
automated coverage.

  ndb_binlog_testevent_rd

Change-Id: Icef0fd0972fb5646bd27dbf2b9ff194938a30ba9