Bug#35663761 NdbApi fails to find container for latest epoch in NODE_FAILREP handling
Background

NdbApi events are buffered in NdbApi to match the rate of production and
rate of consumption by user code.
A maximum buffer size can be set to avoid unbounded memory usage when
the rate is mismatched for a long time.
When the maximum size is reached event buffering stops until the buffer
usage drops below a lower threshold.
A TE_OUT_OF_MEMORY event is added to the buffer to inform the consumer
that there can be missing events.

Problem

Testing and experience shows some issues with this mechanism, particularly as
it stops the buffering of both 'table data' events (TE_INSERT, TE_UPDATE,
TE_DELETE) and other meta events (TE_DROP_TABLE, TE_NODE_FAILURE,
TE_CLUSTER_FAILURE, ...).  Discarding the meta events results in issues
in NdbApi and in consumer code as the state of the event stream is not
clear.

One specific example is where the cluster is disconnected while the event
buffer is full - in this case, the local ClusterMgr thread should add
TE_NODE_FAILURE and TE_CLUSTER_FAILURE events to the buffered events for
each active EventOperation, but this was not possible and resulted in
a crash.  Other undesirable behaviours may be possible.

Solution

In an event buffer overload situation, only the table data events
TE_INSERT, TE_UPDATE, TE_DELETE are discarded.  Meta events are buffered
and eventually processed as normal.

Effects

This means that the buffer usage is not necessarily as tightly bounded
as in the original implementation, though the rate of production of meta
events is expected to be minimal.

Additionally it means that from the consumer's point of view, when they
iterate to a position in the event stream that corresponds to an out
of memory situation, they may still observe some meta events 'during'
the out of memory period.  These events will not include data, and
specifically will not include TE_EMPTY_EPOCH events which could be
misleading.

The consumer can therefore safely assume that the out-of-memory situation
persists until it receives TE_EMPTY_EPOCH (if enabled) or data carrying
events afer a TE_OUT_OF_MEMORY event.

An additional effect of this change is that the event API's latest epoch
as returned by getLatestGCI() no longer is frozen while the event buffer
producer is in an epoch gap - it continues to climb.

This behaviour was not documented before as an indication of event buffer
OOM handling, but was used by some existing test_event testcase.  These
are modified to use observations of the buffer fill level instead to
detect when an OOM gap is being buffered.

Testing

Four new tests added to test_event_mysqld :

  test_event_mysqld -n MySQLDEventsEventBufferOverload
  test_event_mysqld -n MySQLDEventsEventBufferOverloadRestarts
  test_event_mysqld -n MySQLDEventsEventBufferOverloadDisconnects
  test_event_mysqld -n MySQLDEventsEventBufferOverloadRestartsDisconnects

New MTR test added invoking

  test_event_mysqld -n MySQLDEventsEventBufferOverloadRestartsDisconnects

  ndb_binlog_testevent_ord

Change-Id: If904fb60be2798b53cc4cb25ea3607c4289739b9