Bug#35655162 NdbApi event buffer overflow can cause timeout waiting for drop table
Problem

During NdbApi event buffer overflow, buffering of data and metadata
events is paused, resulting in data and metadata events being
discarded.

This means that event subscribers waiting for metadata events
such as TE_DROP_TABLE may never receive them and will timeout.

Specifically, MySQL Cluster schema distribution uses the arrival
of a TE_DROP_TABLE event on an NdbEventOperation to know when it
is required and safe to drop the NdbEventOperation.

If this event does not arrive within a bounded time then the
NdbEventOperation is not dropped, leaking resources and potentially
resulting in later problems.

Solution

The fix to
  Bug#35663761 NdbApi fails to find container for latest epoch in NODE_FAILREP handling

modified the NdbApi event buffer overload handling to only pause
the buffering of data events, but continue buffering metadata
events.
This means that in event buffer overload situations,
TE_DROP events will not be lost.

This patch adds a testcase covering this scenario :

  test_event_mysqld -n MySQLDEventsEventBufferOverloadDDL

This test is invoked from MTR as :
  ndb_binlog_testevent_os

Change-Id: I654fd0aa0c83d35956d4d01632f16f610951bd0c