Bug#35968195 mysql_config --libs reports -lzlib instead of -lz
This patch added some abstaction layers between physical libraries
(bundled or system) and their names on disk vs name in cmake code:
    Bug#35057542 Create INTERFACE libraries for bundled/system zlib/zstd/lz4

When generating the files mysqlclient.pc (for pkg-config) and
mysql_config (a shell script), we need to go through these abstraction
layers, and produce library names as required by the linker of the
host platform. This is done by the cmake macro EXTRACT_LINK_LIBRARIES.

The fix is to store these explicitly as "-lz" and "-lzstd"
respectively.  The lz4 library is not used by our client library, so
there is no need to store the linker option for that.

Change-Id: Ief70f66900bec2d07fdb3daa88caf60a138c953b