Bug#35530823: password from CREATE USER written as plaintext in binlog , breaks replication
CREATE USER IF EXISTS is binlogged no matter if it creates the user or
finds one.
But if it finds one it's bailing out a tad too early and is not getting
the clear text authentication string rewritten to disk format as
a result.
Fixed by setting the (already calculated) in-memory acl string to the
parser statement so that re-writing this in binlog format would actually
get the WITH .. AS variant properly.

Added a test case.
Also made sure the correct password is passed down for rewrite by
invoking the plugin to produce a hash of the supplied password.
Fixed to work on linux by also specifying a socket to the test connect.

Change-Id: I1082086f7686d92c85bedb1e9ac9359a19fafa12