WL#15515: Add TP connection information to PFS
Problem: Inadequate P_S instrumentation for TP

Solution: Add a new P_S plugin table for TP called tp_connections which
displays the information in the tp_client_low_level_t struct for each
connection. In addition the exiting P_S tables tp_thread_group_state and
tp_thread_state are extended with additional columns for existing data structure
members which were missing, and also with new information which will be
useful for performance tuning/diagnostics.

In addition, some limited refactoring was done to simply information
retrieval for, and the modification of the schema for, the TP P_S tables.

Change-Id: I94414afd00b5a6702ee03ec341c9c0b773a4b94c