Bug#35064211: Transaction_monitor_thread::start can stay waiting for the thread to start
Problem:
When the Transaction_monitor_thread is created, the thread that requested for
the creation waits until Transaction_monitor_thread starts running.
However upon running, the Transaction_monitor_thread thread does not unlock the
mutex causing the creator of a thread to wait for a long time to read the thread
running status.

Analysis:
Checked and observed all other threads on running unlocks the mutex.
In Transaction_monitor_thread the lock was never released post setting the
status of the thread as running.

Fix:
Code has benn improved to unlock the mutex upon thread creation after setting
the status of the thread as being running. The scope of the lock has been
reduced.

Change-Id: I07f961346b99a740c76b1e315850921a2f4fcdb8