Bug#35855294 Inconsistent results of SELECT statement with EXISTS clause [noclose]
Normally, subqueries in expressions which do not contain window
functions are executed before the windowing phase.  The results of the
subquery is computed and written to the window's outfile (and also
frame buffer depending on the window specification); the windowing
phase will then only see a field corresponding to the result field of
the subquery when evaluating.

But sometimes, subqueries occur in an expression containing window
functions. Such expressions are mostly [1] evaluated only once the
result of the window function is ready. If such subqueries do not
contain outer refererences, this delay is not an issue.

[1] the one current exception is scalar subqueries, which *are*
evaluated before the window function, which is fine.

However, when subqueries do contain outer references, we have an
issue, as evidenced by the EXISTS repro:

The correlated columns are read from the windowing *input* file during
evaluation, not from the frame buffer/output file. This is in general
wrong, because the buffering of rows in the frame buffer will have
moved past the correct row in the input file.  In the repro, the last
row in the input file has been read so all references to c5 in
expression "BETWEEN ref_0.c5 AND ref_0.c5 refererenced the final row
in the input file, hence the wrong result.

As it turns out, an EXISTS query can be evaluated before windowing,
because it has an empty left_expr, i.e. we do not need the value of
the window function to evaluate the immediate boolean expression which
the subquery constitutes.  The patch moves evaluation of EXISTS
subqueries to before windowing. But this patch is only partial:

If the left side of an ANY/ALL/IN subquery contains a window function,
we will postpone evaluation also and a similar problem as the one
solved above for EXISTS will happen if the subquery is correlated and
the window function requires a frame buffer, e.g. in such a query:

SELECT AVG(f1) OVER (PARTITION BY f1)
  IN (SELECT outer_t.f1 FROM t1) FROM t1 AS outer_t;

With t1 containing (1), (2), the query gives two rows: 0, 1. The
correct result is: 1,1.

This is not fixed by the present patch: we would have to include the
outer columns in the set of columns in the window frame buffer/outfile
(possibly hidden) and to re-write the outer references inside the
subquery to reference the frame buffer/output files.  This will be
fixed in a follow-up patch.

If the left side does *not* contain a window function (as opposed to
only occuring in the same expression

           expr
          /   \
       wf     expr IN subq
              expr > ALL/ANY

we do move it to before the windowing phase (also new in this patch).

Change-Id: I1414646338e3378355d36624d522f80f453aaa09